SenseCom v1.3.0 Release
Fixes the Bluetooth Lag Issue we've been having when one or more device(s) is paired, but turned off. It should stop being an issue after the fist two times trying to connect.
Included new Nova Glove icons to be more representative of the actual device.
Disabled the 'last command sent to the glove' display while we're refactoring some of our haptics back-end.